+++ mixer/test/client/quota/quota_test.go
@@ -28,7 +28,7 @@ const (
 )
 
 // Stats in Envoy proxy.
-var expectedStats = map[string]int{
+var expectedStats = map[string]uint64{
 	// Policy check stats
 	"http_mixer_filter.total_check_calls":             2,
 	"http_mixer_filter.total_check_cache_hits":        0,
